refactor: updating file namas and type for Particles
