--- src/app/(materials)/romajis/romaji-questions.ts
@@ -1,209 +1,0 @@
-export const romajiQuestions: RomajisQuestion[] = [
-  // 促音・長音・清音 (Category 1)
-  {
-    id: 18,
-    category: "romajis",
-    categoryId: 1,
-    group: "sokuon_chouon_seion",
-    groupExp: "いっぱんてきなことば",
-    isIndex: true,
-    hiragana: "がっこう",
-    answer: ["gakkou"],
-  },
-  {
-    id: 19,
-    category: "romajis",
-    categoryId: 2,
-    group: "sokuon_chouon_seion",
-    groupExp: "いっぱんてきなことば",
-    isIndex: false,
-    hiragana: "きって",
-    answer: ["kitte"],
-  },
-  {
-    id: 20,
-    category: "romajis",
-    categoryId: 3,
-    group: "sokuon_chouon_seion",
-    groupExp: "いっぱんてきなことば",
-    isIndex: false,
-    hiragana: "せんせい",
-    answer: ["sensei"],
-  },
-  {
-    id: 21,
-    category: "romajis",
-    categoryId: 4,
-    group: "sokuon_chouon_seion",
-    groupExp: "いっぱんてきなことば",
-    isIndex: false,
-    hiragana: "ほうりつ",
-    answer: ["houritu", "houritsu"],
-  },
-  {
-    id: 22,
-    category: "romajis",
-    categoryId: 5,
-    group: "sokuon_chouon_seion",
-    groupExp: "いっぱんてきなことば",
-    isIndex: false,
-    hiragana: "とうきょう",
-    answer: ["toukyou"],
-  },
-
-  // 撥音 (Category 2)
-  {
-    id: 23,
-    category: "romajis",
-    categoryId: 6,
-    group: "hatsuon",
-    groupExp: "「ん」をふくむことば",
-    isIndex: true,
-    hiragana: "えんぴつ",
-    answer: ["enpitu", "enpitsu"],
-  },
-  {
-    id: 24,
-    category: "romajis",
-    categoryId: 7,
-    group: "hatsuon",
-    groupExp: "「ん」をふくむことば",
-    isIndex: false,
-    hiragana: "しんぶん",
-    answer: ["shinbun", "sinbun"],
-  },
-  {
-    id: 25,
-    category: "romajis",
-    categoryId: 8,
-    group: "hatsuon",
-    groupExp: "「ん」をふくむことば",
-    isIndex: false,
-    hiragana: "けんこう",
-    answer: ["kenkou"],
-  },
-  {
-    id: 26,
-    category: "romajis",
-    categoryId: 9,
-    group: "hatsuon",
-    groupExp: "「ん」をふくむことば",
-    isIndex: false,
-    hiragana: "はんたい",
-    answer: ["hantai"],
-  },
-  {
-    id: 27,
-    category: "romajis",
-    categoryId: 10,
-    group: "hatsuon",
-    groupExp: "「ん」をふくむことば",
-    isIndex: false,
-    hiragana: "おんがく",
-    answer: ["ongaku"],
-  },
-
-  // 拗音 (Category 3)
-  {
-    id: 28,
-    category: "romajis",
-    categoryId: 11,
-    group: "youon",
-    groupExp: "「ゃ、ゅ、ょ」をふくむことば",
-    isIndex: true,
-    hiragana: "おきゃくさま",
-    answer: ["okyakusama"],
-  },
-  {
-    id: 29,
-    category: "romajis",
-    categoryId: 12,
-    group: "youon",
-    groupExp: "「ゃ、ゅ、ょ」をふくむことば",
-    isIndex: false,
-    hiragana: "しゅくだい",
-    answer: ["shukudai", "syukudai"],
-  },
-  {
-    id: 30,
-    category: "romajis",
-    categoryId: 13,
-    group: "youon",
-    groupExp: "「ゃ、ゅ、ょ」をふくむことば",
-    isIndex: false,
-    hiragana: "じゅぎょう",
-    answer: ["jugyou", "jyugyou", "zyugyou"],
-  },
-  {
-    id: 31,
-    category: "romajis",
-    categoryId: 14,
-    group: "youon",
-    groupExp: "「ゃ、ゅ、ょ」をふくむことば",
-    isIndex: false,
-    hiragana: "ちゃいろ",
-    answer: ["chairo", "cyairo", "tyairo"],
-  },
-  {
-    id: 32,
-    category: "romajis",
-    categoryId: 15,
-    group: "youon",
-    groupExp: "「ゃ、ゅ、ょ」をふくむことば",
-    isIndex: false,
-    hiragana: "りょこう",
-    answer: ["ryokou"],
-  },
-
-  // 濁音・半濁音 (Category 4)
-  {
-    id: 33,
-    category: "romajis",
-    categoryId: 16,
-    group: "dakuon",
-    groupExp: "にごる発音をふくむことば",
-    isIndex: true,
-    hiragana: "じどうしゃ",
-    answer: ["jidousha", "jidousya", "zidousya"],
-  },
-  {
-    id: 34,
-    category: "romajis",
-    categoryId: 17,
-    group: "dakuon",
-    groupExp: "にごる発音をふくむことば",
-    isIndex: false,
-    hiragana: "でんしゃ",
-    answer: ["densha", "densya"],
-  },
-  {
-    id: 35,
-    category: "romajis",
-    categoryId: 18,
-    group: "dakuon",
-    groupExp: "にごる発音をふくむことば",
-    isIndex: false,
-    hiragana: "びょういん",
-    answer: ["byouin"],
-  },
-  {
-    id: 36,
-    category: "romajis",
-    categoryId: 19,
-    group: "dakuon",
-    groupExp: "にごる発音をふくむことば",
-    isIndex: false,
-    hiragana: "げんき",
-    answer: ["genki"],
-  },
-  {
-    id: 37,
-    category: "romajis",
-    categoryId: 20,
-    group: "dakuon",
-    groupExp: "にごる発音をふくむことば",
-    isIndex: false,
-    hiragana: "ぴあの",
-    answer: ["piano"],
-  },
-];
