+++ src/components/menu-board/particles-menu-board.component.tsx
@@ -1,14 +1,18 @@
 "use client";
 
-import PaticlesMenuBtn from "../menu-btns/particles-menu-btn.component";
+import PaticlesMenuBtn from "../menu-btn/particles-menu-btn.component";
 
-export default function ParticlesMenuBoard({ params }: { params: Question[] }) {
+export default function ParticlesMenuBoard({
+  params,
+}: {
+  params: ParticlesQuestion[];
+}) {
   return (
     <div className="h-full flex flex-col justify-center">
       <div className="flex flex-row-reverse justify-center gap-10">
         {params
-          .filter((item: Question) => item.isIndex)
-          .map((item: Question) => {
+          .filter((item: ParticlesQuestion) => item.isIndex)
+          .map((item: ParticlesQuestion) => {
             return (
               <div key={item.id}>
                 <PaticlesMenuBtn params={item} />
